requirements.txt
streamlit==1.41.1
openai==1.59.9
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
faiss-cpu==1.9.0.post1
PyPDF2==3.0.1
python-dotenv==1.0.1
tiktoken==0.8.0
python-docx==1.1.2
